pyproject.toml
[project]
name = "slick"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "google-genai>=1.68.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
